+++ implement/Pine.Core/LanguageServerProtocol/RenameParams.cs
@@ -1,0 +1,10 @@
+namespace Pine.Core.LanguageServerProtocol;
+
+/// <summary>
+/// https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#renameParams
+/// </summary>
+public record RenameParams(
+    TextDocumentIdentifier TextDocument,
+    Position Position,
+    string NewName);
+