+++ implement/Pine.Core/LanguageServerProtocol/TextDocumentEdit.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Pine.Core.LanguageServerProtocol;
+
+/// <summary>
+/// https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocumentEdit
+/// </summary>
+public record TextDocumentEdit(
+    OptionalVersionedTextDocumentIdentifier TextDocument,
+    IReadOnlyList<TextEdit> Edits);
+